src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
scipy==1.11.4
scikit-learn==1.3.2
